requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
faiss-cpu==1.11.0.post1
sentence-transformers==5.0.0
langchain==0.3.26
langchain-core==0.3.72
langchain-community==0.3.27
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.8
google-generativeai==0.8.5
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.3.0
streamlit==1.47.0
altair==5.5.0
watchdog==6.0.0
python-dotenv==1.1.1
regex==2024.11.6
requests==2.32.4
tqdm==4.67.1
pandas==2.3.1
numpy==2.3.1
typing_extensions==4.14.1
Jinja2==3.1.6
